+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `EmptyState` to `EXPERIMENTAL_TableV2`
+
 ## [9.78.10] - 2019-09-10
 
 ### Fixed